Port several platforms from `llvm_asm` to `asm` where the change is trivial
The reason for this change is that `llvm_asm` is removed from nightly, so it
doesn't compile anymore on any current compiler, while `asm` is being stabilised
(currently stable on beta).

The major omission in this commit is the linux x86 (32 bit) platform. The new `asm`
macro doesn't allow the esi and ebp registers to be used as inputs or outputs or
clobbers, which makes the port non-trivial for this platform. Similarly on armeabi
the r6 register is not allowed as input, output or clobber.